requirements.txt
torch
torchvision
Pillow
numpy
cv2
openai>=1.0.0
requests>=2.28.0